Nouveau composant <RegionSelector
